Add files via upload
Caeser cipher is a mathematical process used in encoding messages. It does this by the application of subtraction where alphabets in a word or sentence (mostly English) are shifted by fixed number of spaces to arrive at an encoded message. The encoded message can be decoded by shifting with the same number of spaces employed in encoding the message except that addition is applied in place of subtraction. Try out the code I wrote for the implementation of this application.